[Template] migrates system templates

+++ src/main/core/Installation/Migrations/Version20250226171749.php
@@ -2,6 +2,7 @@
 
 namespace Claroline\CoreBundle\Installation\Migrations;
 
+use Claroline\InstallationBundle\Migrations\Helper\ConditionalMigrationTrait;
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;
 
@@ -12,6 +13,8 @@ use Doctrine\Migrations\AbstractMigration;
  */
 final class Version20250226171749 extends AbstractMigration
 {
+    use ConditionalMigrationTrait;
+
     public function up(Schema $schema): void
     {
         $this->addSql('
@@ -19,11 +22,13 @@ final class Version20250226171749 extends AbstractMigration
             ADD downloadable TINYINT(1) DEFAULT 0 NOT NULL
         ');
 
-        // there is no cascade on slides, and it will break on the plugin removal
-        // if the table is not empty
-        $this->addSql('
-            DELETE FROM claro_slide
-        ');
+        if ($this->checkTableExists('claro_slide', $this->connection)) {
+            // there is no cascade on slides, and it will break on the plugin removal
+            // if the table is not empty
+            $this->addSql('
+                DELETE FROM claro_slide
+            ');
+        }
     }
 
     public function down(Schema $schema): void
